refactor(decompose): split electron-real-folder.spec.ts into folder
Extracts fixtures, helpers, types, fs-helpers, ui-helpers, and test-data
into a sibling electron-real-folder/ folder. The .spec.ts entrypoint stays
in place (playwright glob discovery) and shrinks from 1370 to ~291 lines.
Behavior preserved: no test names, ordering, selectors, or assertions
changed.
