+++ webapp/e2e-tests/electron/for_feature_development_not_LT_verification/graph_folder_loading/electron-real-folder.spec.ts
@@ -10,421 +10,62 @@
  * IMPORTANT: THESE SPEC COMMENTS MUST BE KEPT UP TO DATE
  */
 
-import { test as base, expect, _electron as electron } from '@playwright/test';
-import type { ElectronApplication, Page } from '@playwright/test';
-import * as path from 'path';
-import * as fs from 'fs/promises';
-import * as os from 'os';
-import type { Core as CytoscapeCore, NodeSingular, EdgeSingular } from 'cytoscape';
-import type { ElectronAPI } from '@/shell/electron';
-
-// Use absolute paths for example_folder_fixtures
-const PROJECT_ROOT = path.resolve(process.cwd());
-const FIXTURE_VAULT_PATH = path.join(PROJECT_ROOT, 'example_folder_fixtures', 'example_real_large', '2025-09-30');
-
-// Type definitions
-interface ExtendedWindow {
-  cytoscapeInstance?: CytoscapeCore;
-  electronAPI?: ElectronAPI;
-  testHelpers?: {
-    createTerminal: (nodeId: string) => void;
-    addNodeAtPosition: (position: { x: number; y: number }) => Promise<void>;
-    getEditorInstance: (windowId: string) => { getValue: () => string; setValue: (content: string) => void } | undefined;
-  };
-}
-
-interface GraphState {
-  nodeCount: number;
-  edgeCount: number;
-  nodeLabels: string[];
-  edges: Array<{ source: string; target: string }>;
-}
-
-
-// Extend test with Electron app
-const test = base.extend<{
-  electronApp: ElectronApplication;
-  appWindow: Page;
-}>({
-  // Set up Electron application
-
-  electronApp: async ({}, use) => {
-    // Create a temporary userData directory for this test (like smoke test does)
-    const tempUserDataPath = await fs.mkdtemp(path.join(os.tmpdir(), 'voicetree-real-folder-test-'));
-
-    // Write the config file to auto-load the test vault (like smoke test does)
-    // IMPORTANT: Set empty suffix so it uses the directory directly, not directory/voicetree
-    const configPath = path.join(tempUserDataPath, 'voicetree-config.json');
-    await fs.writeFile(configPath, JSON.stringify({
-      lastDirectory: FIXTURE_VAULT_PATH,
-      suffixes: {
-        [FIXTURE_VAULT_PATH]: '' // Empty suffix means use directory directly
-      }
-    }, null, 2), 'utf8');
-    console.log('[Real Folder Test] Created config file to auto-load:', FIXTURE_VAULT_PATH);
-
-    const electronApp = await electron.launch({
-      args: [
-        path.join(PROJECT_ROOT, 'dist-electron/main/index.js'),
-        `--user-data-dir=${tempUserDataPath}` // Use temp userData to isolate test config
-      ],
-      env: {
-        ...process.env,
-        NODE_ENV: 'test',
-        HEADLESS_TEST: '1',
-        MINIMIZE_TEST: '1' // Minimize window to avoid dialog popups
-      },
-      timeout: 10000 // Standard timeout for app launch
-    });
-
-    await use(electronApp);
-
-    // Graceful shutdown: Stop file watching before closing app
-    // This prevents EPIPE errors from file watcher trying to log after stdout closes
-    try {
-      const window = await electronApp.firstWindow();
-      await window.evaluate(async () => {
-        const api = (window as unknown as ExtendedWindow).electronAPI;
-        if (api) {
-          await api.main.stopFileWatching();
-        }
-      });
-      // Wait for pending file system events to drain
-      await window.waitForTimeout(300);
-    } catch {
-      // Window might already be closed, that's okay
-      console.log('Note: Could not stop file watching during cleanup (window may be closed)');
-    }
-
-    await electronApp.close();
-    console.log('[Real Folder Test] Electron app closed');
-
-    // Cleanup temp directory
-    await fs.rm(tempUserDataPath, { recursive: true, force: true });
-  },
-
-  // Get the main window
-  appWindow: async ({ electronApp }, use) => {
-    const window = await electronApp.firstWindow({ timeout: 10000 });
-
-    // Log console messages for debugging
-    window.on('console', msg => {
-      console.log(`BROWSER [${msg.type()}]:`, msg.text());
-    });
-
-    // Capture page errors
-    window.on('pageerror', error => {
-      console.error('PAGE ERROR:', error.message);
-      console.error('Stack:', error.stack);
-    });
-
-    await window.waitForLoadState('domcontentloaded');
-
-    // Check for errors before waiting for cytoscapeInstance
-    const hasErrors = await window.evaluate(() => {
-      const errors: string[] = [];
-      // Check if React rendered
-      if (!document.querySelector('#root')) errors.push('No #root element');
-      // Check if any error boundaries triggered
-      const errorText = document.body.textContent;
-      if (errorText?.includes('Error') || errorText?.includes('error')) {
-        errors.push(`Page contains error text: ${errorText.substring(0, 200)}`);
-      }
-      return errors;
-    });
-
-    if (hasErrors.length > 0) {
-      console.error('Pre-initialization errors:', hasErrors);
-    }
-
-    await window.waitForFunction(() => (window as unknown as ExtendedWindow).cytoscapeInstance, { timeout: 10000 });
-
-    // Wait a bit longer to ensure graph is ready
-    await window.waitForTimeout(1000);
-
-    await use(window);
-  }
-});
+import { expect, test } from './electron-real-folder/fixtures';
+import * as realFolder from './electron-real-folder/helpers';
+import * as vault from './electron-real-folder/fs-helpers';
+import {
+  COMPLEX_LINKS_CONTENT,
+  INCREMENTAL_TEST_FILES,
+  NEW_CONCEPT_CONTENT
+} from './electron-real-folder/test-data';
+import type { GraphState, ViewportState } from './electron-real-folder/types';
 
 test.describe('Real Folder E2E Tests', () => {
-  // Cleanup hook to ensure test files are removed even if test fails
-  test.afterEach(async ({ appWindow }) => {
-    // Stop file watching BEFORE cleaning up files to prevent EPIPE errors
-    try {
-      await appWindow.evaluate(async () => {
-        const api = (window as ExtendedWindow).electronAPI;
-        if (api) {
-          await api.main.stopFileWatching();
-        }
-      });
-      // Brief wait to let file watcher fully stop
-      await appWindow.waitForTimeout(200);
-    } catch {
-      // Window might be closed, that's okay
-    }
-
-    const testFilesToCleanup = [
-      'incremental-test-1.md',
-      'incremental-test-2.md',
-      'incremental-test-3.md'
-    ];
-
-    for (const fileName of testFilesToCleanup) {
-      const filePath = path.join(FIXTURE_VAULT_PATH, fileName);
-      try {
-        await fs.unlink(filePath);
-        console.log(`Cleaned up leftover file: ${fileName}`);
-      } catch {
-        // File doesn't exist, which is fine
-      }
-    }
-  });
-
   test('should load and visualize a real markdown vault', async ({ appWindow }) => {
-    console.log('=== STEP 1: Verify app loaded ===');
-
-    // Verify app loaded properly (already waited in fixture)
-    const appReady = await appWindow.evaluate(() => {
-      return !!(window as ExtendedWindow).cytoscapeInstance &&
-             !!(window as ExtendedWindow).electronAPI;
-    });
+    const appReady = await realFolder.isAppReady(appWindow);
     expect(appReady).toBe(true);
-    console.log('✓ App loaded successfully');
-
-    console.log('=== STEP 2: Wait for auto-load from config (no manual startFileWatching needed) ===');
-    console.log(`Vault should auto-load from config: ${FIXTURE_VAULT_PATH}`);
-
-    // Wait for auto-load to complete and nodes to appear in graph
-    // The app auto-loads from config (like smoke test), so nodes should appear automatically
-    await expect.poll(async () => {
-      return appWindow.evaluate(() => {
-        const cy = (window as ExtendedWindow).cytoscapeInstance;
-        if (!cy) return 0;
-        return cy.nodes().length;
-      });
-    }, {
+    await expect.poll(() => realFolder.getNodeCount(appWindow), {
       message: 'Waiting for graph to load nodes',
       timeout: 15000,
       intervals: [500, 1000, 1000]
     }).toBeGreaterThan(0);
-
-    console.log('✓ Graph auto-loaded from config');
-
-    console.log('=== STEP 3: Verify initial graph state ===');
-
-    // Get the initial graph state
-    const initialGraph: GraphState = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not initialized');
-
-      return {
-        nodeCount: cy.nodes().length,
-        edgeCount: cy.edges().length,
-        nodeLabels: cy.nodes().map((n: NodeSingular) => n.data('label')).sort(),
-        edges: cy.edges().map((e: EdgeSingular) => ({
-          source: e.source().data('label'),
-          target: e.target().data('label')
-        }))
-      };
-    });
-
-    console.log(`Graph state: ${initialGraph.nodeCount} nodes, ${initialGraph.edgeCount} edges`);
-    console.log('GraphNode labels:', initialGraph.nodeLabels);
-
-    // Verify expected files are loaded
-    // Note: Labels are extracted from frontmatter title > first heading > filename
+    const initialGraph: GraphState = await realFolder.getGraphState(appWindow);
     expect(initialGraph.nodeCount).toBeGreaterThanOrEqual(5); // Fixture has 56 files
     expect(initialGraph.nodeLabels).toContain('Setting up Agent in Feedback Loop'); // From heading in file 10
     expect(initialGraph.nodeLabels).toContain('Identify Relevant Test'); // From heading in file 11
-
-    // Verify outgoingEdges exist (wiki-links create outgoingEdges)
     expect(initialGraph.edgeCount).toBeGreaterThan(0);
-
-    // BEHAVIORAL TEST: Verify outgoingEdges are actually visible (not just present in data)
-    const edgeVisibility = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not initialized');
-
-      // Filter out ghost outgoingEdges (invisible outgoingEdges to GHOST_ROOT_ID)
-      const visibleEdges = cy.edges('[!isGhostEdge]');
-      if (visibleEdges.length === 0) return { visible: false, reason: 'no visible outgoingEdges' };
-
-      // Sample first visible edge to check visibility styles
-      const edge = visibleEdges.first();
-      const opacity = parseFloat(edge.style('opacity'));
-      const width = parseFloat(edge.style('width'));
-      const color = edge.style('line-color');
-
-      return {
-        visible: opacity > 0 && width > 0 && color !== 'transparent',
-        opacity,
-        width,
-        color,
-        edgeCount: visibleEdges.length
-      };
-    });
-
-    console.log('Edge visibility check:', edgeVisibility);
+    const edgeVisibility = await realFolder.getEdgeVisibility(appWindow);
     expect(edgeVisibility.visible).toBe(true);
     expect(edgeVisibility.opacity).toBeGreaterThan(0);
     expect(edgeVisibility.width).toBeGreaterThan(0);
-
-    // BEHAVIORAL TEST: Verify outgoingEdges have labels displayed
-    const edgeLabelCheck = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not initialized');
-
-      const edges = cy.edges();
-      if (edges.length === 0) return { hasLabels: false, reason: 'no outgoingEdges' };
-
-      // Sample all outgoingEdges to check if ANY have labels
-      const edgesWithLabels = edges.filter(e => {
-        const label = e.data('label');
-        return label && label.length > 0;
-      });
-
-      // Get style for edge labels
-      const firstEdge = edges.first();
-      const labelText = firstEdge.style('label');
-      const fontSize = firstEdge.style('font-size');
-
-      return {
-        totalEdges: edges.length,
-        edgesWithDataLabels: edgesWithLabels.length,
-        sampleLabel: edges.first().data('label'),
-        styleLabelValue: labelText,
-        fontSize: fontSize
-      };
-    });
-
-    console.log('Edge label check (initial):', edgeLabelCheck);
-    // Initial outgoingEdges from plain [[wikilinks]] won't have labels
+    const edgeLabelCheck = await realFolder.getEdgeLabelCheck(appWindow);
     expect(edgeLabelCheck.totalEdges).toBeGreaterThan(0);
-
-    // SKIPPED: Edge labels with relationship types (e.g., "references", "extends")
-    // The current functional graph system doesn't extract relationship types from markdown.
-    // Edges are simple connections without labels. This is a future feature.
-    console.log('✓ Edge label extraction not yet implemented (future feature)');
-    console.log('✓ Initial graph loaded correctly');
-
-    console.log('=== STEP 4: Test file modification ===');
-
-    // Capture node count right before adding new file (after cleanup)
-    const nodeCountBeforeAdd = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) return 0;
-      return cy.nodes().length;
-    });
-    console.log(`Node count before adding new-concept: ${nodeCountBeforeAdd}`);
-
-    // Create a new file in the vault with links to existing files
-    const newFilePath = path.join(FIXTURE_VAULT_PATH, 'new-concept.md');
-    await fs.writeFile(newFilePath, `# New Concept
-
-This is a dynamically added concept that links to [[10_Setting_up_Agent_in_Feedback_Loop]] and [[14_Assign_Agent_to_Identify_Boundaries]].
-
-It demonstrates that the file watcher detects new files in real-time.`);
-
-    // Wait for the new file to be detected and processed
-    await expect.poll(async () => {
-      return appWindow.evaluate(() => {
-        const cy = (window as ExtendedWindow).cytoscapeInstance;
-        if (!cy) return false;
-
-        const labels = cy.nodes().map((n: NodeSingular) => n.data('label'));
-        return labels.includes('New Concept'); // Label from heading
-      });
-    }, {
+    const nodeCountBeforeAdd = await realFolder.getNodeCount(appWindow);
+    const newFilePath = await vault.writeVaultFile('new-concept.md', NEW_CONCEPT_CONTENT);
+    await expect.poll(() => realFolder.hasNodeLabel(appWindow, 'New Concept'), {
       message: 'Waiting for new-concept node to appear',
       timeout: 10000
     }).toBe(true);
-
-    // Verify the graph updated correctly
-    const updatedGraph: GraphState = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not available');
-
-      return {
-        nodeCount: cy.nodes().length,
-        edgeCount: cy.edges().length,
-        nodeLabels: cy.nodes().map((n: NodeSingular) => n.data('label')).sort(),
-        edges: cy.edges().map((e: EdgeSingular) => ({
-          source: e.source().data('label'),
-          target: e.target().data('label')
-        }))
-      };
-    });
-
-    console.log(`Node count after adding new-concept: ${updatedGraph.nodeCount} (expected ${nodeCountBeforeAdd + 1})`);
-    console.log('GraphNode labels:', updatedGraph.nodeLabels);
-
-    // Verify new-concept node exists (using label from heading)
+    const updatedGraph: GraphState = await realFolder.getGraphStateFromAvailableCytoscape(appWindow);
     expect(updatedGraph.nodeLabels).toContain('New Concept');
-
-    // GraphNode count should have increased (allowing for possible placeholder cleanup)
     expect(updatedGraph.nodeCount).toBeGreaterThanOrEqual(nodeCountBeforeAdd);
-
-    // Check that outgoingEdges were created for the wiki-links (using labels from frontmatter/headings)
     const newConceptEdges = updatedGraph.edges.filter(e =>
       e.source === 'New Concept' || e.target === 'New Concept'
     );
     expect(newConceptEdges.length).toBeGreaterThan(0);
-    console.log('✓ File addition detected and graph updated');
-
-    console.log('=== STEP 5: Test file deletion ===');
-
-    // Delete the file we just created
-    await fs.unlink(newFilePath);
-
-    // Wait for the file deletion to be processed
-    await expect.poll(async () => {
-      return appWindow.evaluate(() => {
-        const cy = (window as ExtendedWindow).cytoscapeInstance;
-        if (!cy) return true; // Still processing
-
-        const labels = cy.nodes().map((n: NodeSingular) => n.data('label'));
-        return !labels.includes('New Concept'); // Label from heading
-      });
-    }, {
+    await vault.deleteFilePath(newFilePath);
+    await expect.poll(() => realFolder.lacksNodeLabel(appWindow, 'New Concept'), {
       message: 'Waiting for new-concept node to be removed',
       timeout: 10000
     }).toBe(true);
-
-    // Verify we're back to the state before adding new-concept
-    const finalGraph: GraphState = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not available');
-
-      return {
-        nodeCount: cy.nodes().length,
-        edgeCount: cy.edges().length,
-        nodeLabels: cy.nodes().map((n: NodeSingular) => n.data('label')).sort(),
-        edges: cy.edges().map((e: EdgeSingular) => ({
-          source: e.source().data('label'),
-          target: e.target().data('label')
-        }))
-      };
-    });
-
-    // Node count should be close to original (within 1-2 nodes due to potential placeholder nodes)
+    const finalGraph: GraphState = await realFolder.getGraphStateFromAvailableCytoscape(appWindow);
     expect(finalGraph.nodeCount).toBeGreaterThanOrEqual(nodeCountBeforeAdd - 2);
     expect(finalGraph.nodeCount).toBeLessThanOrEqual(nodeCountBeforeAdd + 2);
     expect(finalGraph.nodeLabels).not.toContain('New Concept');
-    console.log('✓ File deletion detected and graph updated');
-
-    console.log('=== STEP 6: Verify wiki-link relationships ===');
-
-    // Check for wiki-link edges that reliably exist from fixture files
-    // Note: Labels come from first heading > first non-empty line > filename (NOT frontmatter title)
-
-    // 14_Assign_Agent_to_Identify_Boundaries.md has heading "### Action to assign an agent..."
-    // This node should have incoming edges from other files that link to it
     const assignNodeLabel: string = 'Action to assign an agent to identify code extraction boundaries.';
     const hasEdgesToAssignNode: boolean = finalGraph.edges.some(e =>
       e.target === assignNodeLabel
     );
-
-    // 17_Create_G_Cloud_Configuration.md or 4_Setup_G_Cloud_CLI.md should be linked
     const hasCloudConfigOrSetupLink: boolean = finalGraph.edges.some(e =>
       e.source.includes('G Cloud Configuration') ||
       e.target.includes('G Cloud CLI')
@@ -432,940 +73,219 @@ It demonstrates that the file watcher detects new files in real-time.`);
 
     expect(hasEdgesToAssignNode).toBe(true);
     expect(hasCloudConfigOrSetupLink).toBe(true);
-    console.log('✓ Wiki-link relationships correctly represented');
-
-    console.log('=== STEP 7: Stop file watching ===');
-
-    // Stop watching
-    const stopResult = await appWindow.evaluate(async () => {
-      const api = (window as ExtendedWindow).electronAPI;
-      if (!api) throw new Error('electronAPI not available');
-
-      return await api.main.stopFileWatching();
-    });
+    const stopResult = await realFolder.stopFileWatching(appWindow);
 
     expect(stopResult.success).toBe(true);
-
-    // Note: Stopping file watching no longer clears the graph automatically
-    // The graph persists after stopping file watching, which is expected behavior
-    console.log('✓ File watching stopped');
-    console.log('\n✅ Real folder E2E test completed successfully!');
   });
 
   test('should handle complex wiki-link patterns', async ({ appWindow }) => {
-    console.log('=== Testing complex wiki-link patterns ===');
-
-    // Wait for auto-load to complete (vault loads from config automatically)
-    await expect.poll(async () => {
-      return appWindow.evaluate(() => {
-        const cy = (window as ExtendedWindow).cytoscapeInstance;
-        if (!cy) return 0;
-        return cy.nodes().length;
-      });
-    }, {
+    await expect.poll(() => realFolder.getNodeCount(appWindow), {
       message: 'Waiting for graph to load nodes',
       timeout: 15000,
       intervals: [500, 1000, 1000]
     }).toBeGreaterThan(0);
-
-    // Create a file with various wiki-link formats (linking to actual fixture files)
-    const complexLinkFile = path.join(FIXTURE_VAULT_PATH, 'complex-links.md');
-    await fs.writeFile(complexLinkFile, `# Complex Links Test
-
-## Different Link Formats
-- Basic: [[10_Setting_up_Agent_in_Feedback_Loop]]
-- Another: [[14_Assign_Agent_to_Identify_Boundaries]]
-- Non-existent: [[ghost-file]]
-- Self-reference: [[complex-links]]
-
-## Multiple Links in One Line
-Check out [[17_Create_G_Cloud_Configuration]], [[16_Resolve_G_Cloud_CLI_MFA_Block]], and [[14_1_Victor_Append_Agent_Extraction_Analysis_Complete]] for more info.
-
-## Links in Lists
-1. First point about [[20_Create_Cloud_Functions_Directory]]
-2. Second point referencing [[19_Bare_Minimum_Conversion_Requirements_for_Quick_Test]]
-3. Third point linking to [[18_Observation_Plan_Complexity_for_Agent_Conversion]]`);
-
-    // Wait for the file to be processed
-    await expect.poll(async () => {
-      return appWindow.evaluate(() => {
-        const cy = (window as ExtendedWindow).cytoscapeInstance;
-        if (!cy) return false;
-        const labels = cy.nodes().map((n: NodeSingular) => n.data('label'));
-        return labels.includes('Complex Links Test'); // Label from heading
-      });
-    }, {
+    const complexLinkFile = await vault.writeVaultFile('complex-links.md', COMPLEX_LINKS_CONTENT);
+    await expect.poll(() => realFolder.hasNodeLabel(appWindow, 'Complex Links Test'), {
       message: 'Waiting for complex-links file to be processed',
       timeout: 10000
     }).toBe(true);
-
-    // Verify the complex links created appropriate outgoingEdges
-    const graphWithComplexLinks = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not available');
-
-      const complexNode = cy.nodes().filter((n: NodeSingular) =>
-        n.data('label') === 'Complex Links Test' // Label from heading
-      );
-
-      if (complexNode.length === 0) return null;
-
-      const connectedEdges = cy.edges().filter((e: EdgeSingular) =>
-        e.source().id() === complexNode[0].id() ||
-        e.target().id() === complexNode[0].id()
-      );
-
-      return {
-        nodeExists: true,
-        connectedEdgeCount: connectedEdges.length,
-        connections: connectedEdges.map((e: EdgeSingular) => ({
-          source: e.source().data('label'),
-          target: e.target().data('label')
-        }))
-      };
-    });
+    const graphWithComplexLinks = await realFolder.getComplexLinksGraphState(appWindow);
 
     expect(graphWithComplexLinks).not.toBeNull();
     expect(graphWithComplexLinks!.nodeExists).toBe(true);
-
-    // Should have connections to multiple files mentioned in the content
     expect(graphWithComplexLinks!.connectedEdgeCount).toBeGreaterThan(2);
-
-    console.log(`✓ Complex links created ${graphWithComplexLinks!.connectedEdgeCount} edges`);
-    console.log('Connections:', graphWithComplexLinks!.connections);
-
-    // Clean up
-    await fs.unlink(complexLinkFile);
-
-    // Wait for file deletion to be processed
+    await vault.deleteFilePath(complexLinkFile);
     await appWindow.waitForTimeout(500);
-
-    // Stop file watching before test ends
-    const stopResult = await appWindow.evaluate(async () => {
-      const api = (window as ExtendedWindow).electronAPI;
-      if (!api) throw new Error('electronAPI not available');
-      return await api.main.stopFileWatching();
-    });
+    const stopResult = await realFolder.stopFileWatching(appWindow);
 
     expect(stopResult.success).toBe(true);
-    console.log('✓ Complex wiki-link patterns test completed');
   });
 
   test('should bulk load then incrementally add nodes with proper layout', async ({ appWindow }) => {
-    console.log('=== Testing Bulk Load + Incremental Layout (Production Flow) ===');
-
-    // Clean up any leftover test files from previous runs
-    const testFileNames = ['incremental-test-1.md', 'incremental-test-2.md', 'incremental-test-3.md'];
-    for (const fileName of testFileNames) {
-      const filePath = path.join(FIXTURE_VAULT_PATH, fileName);
-      try {
-        await fs.unlink(filePath);
-        console.log(`Cleaned up leftover file: ${fileName}`);
-      } catch {
-        // File doesn't exist, which is fine
-      }
-    }
-
-    // Wait for auto-load to complete (vault loads from config automatically)
-    await expect.poll(async () => {
-      return appWindow.evaluate(() => {
-        const cy = (window as ExtendedWindow).cytoscapeInstance;
-        if (!cy) return 0;
-        return cy.nodes().length;
-      });
-    }, {
+    const testFileNames = vault.INCREMENTAL_TEST_FILE_NAMES;
+    await vault.deleteVaultFilesIfPresent(testFileNames);
+    await expect.poll(() => realFolder.getNodeCount(appWindow), {
       message: 'Waiting for bulk load to complete',
       timeout: 15000,
       intervals: [500, 1000, 1000]
     }).toBeGreaterThan(0);
-
-    // Wait for layout to be applied (auto-layout has 300ms debounce + layout time)
     await appWindow.waitForTimeout(2000);
 
-    console.log('=== PHASE 1: Verify Bulk Load Layout Quality ===');
-
-    const bulkLoadState = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not initialized');
-
-      const nodes = cy.nodes();
-      const positions = nodes.map(n => ({
-        id: n.id(),
-        x: n.position().x,
-        y: n.position().y
-      }));
-
-      // Check Y-coordinate distribution
-      const yCoords = positions.map(p => p.y);
-      const uniqueY = new Set(yCoords);
-      const allAtZero = yCoords.every(y => y === 0);
-
-      return {
-        nodeCount: nodes.length,
-        yCoords,
-        uniqueYCount: uniqueY.size,
-        allAtZero,
-        samplePositions: positions.slice(0, 5)
-      };
-    });
-
-    console.log(`Bulk load: ${bulkLoadState.nodeCount} nodes`);
-    console.log(`Y-coordinates: ${bulkLoadState.uniqueYCount} unique levels`);
-    console.log(`Sample positions:`, bulkLoadState.samplePositions);
-
-    // Critical check: ensure bulk layout worked (not all at y=0)
+    const bulkLoadState = await realFolder.getBulkLoadState(appWindow);
     expect(bulkLoadState.allAtZero).toBe(false);
     expect(bulkLoadState.uniqueYCount).toBeGreaterThan(1);
-    console.log('✓ Bulk load layout has proper Y-coordinate distribution');
 
-    console.log('=== PHASE 2: Add 3 New Nodes Incrementally ===');
-
-    const newFiles = [
-      {
-        name: 'incremental-test-1.md',
-        content: '# Incremental Test 1\n\nFirst incrementally added node. Links to [[introduction]].'
-      },
-      {
-        name: 'incremental-test-2.md',
-        content: '# Incremental Test 2\n\nSecond incremental node. References [[architecture]] and [[incremental-test-1]].'
-      },
-      {
-        name: 'incremental-test-3.md',
-        content: '# Incremental Test 3\n\nThird incremental node. Connects to [[core-principles]].'
-      }
-    ];
-
-    // Add files one by one and verify layout updates
+    const newFiles = INCREMENTAL_TEST_FILES;
     for (let i = 0; i < newFiles.length; i++) {
       const file = newFiles[i];
-      const filePath = path.join(FIXTURE_VAULT_PATH, file.name);
-
-      console.log(`Adding file ${i + 1}: ${file.name}`);
-      await fs.writeFile(filePath, file.content);
-
-      // Wait for file to be detected and laid out
-      await expect.poll(async () => {
-        return appWindow.evaluate((index) => {
-          const cy = (window as ExtendedWindow).cytoscapeInstance;
-          if (!cy) return false;
-          const labels = cy.nodes().map((n: NodeSingular) => n.data('label'));
-          const expectedLabel = `Incremental Test ${index + 1}`; // From heading
-          return labels.includes(expectedLabel);
-        }, i);
-      }, {
+      await vault.writeVaultFile(file.name, file.content);
+      await expect.poll(() => realFolder.hasIncrementalNode(appWindow, i), {
         message: `Waiting for ${file.name} to appear in graph`,
         timeout: 10000
       }).toBe(true);
-
-      console.log(`✓ File ${i + 1} added and detected`);
     }
 
-    console.log('=== PHASE 3: Verify Incremental Layout Quality ===');
-
-    const finalState = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not initialized');
-
-      const nodes = cy.nodes();
-      const positions = nodes.map(n => ({
-        id: n.id(),
-        x: n.position().x,
-        y: n.position().y
-      }));
-
-      // Get the 3 new nodes specifically (by label, which comes from heading)
-      const newNodeLabels = [
-        'Incremental Test 1',
-        'Incremental Test 2',
-        'Incremental Test 3'
-      ];
-
-      const newNodes = cy.nodes().filter((n: NodeSingular) =>
-        newNodeLabels.includes(n.data('label'))
-      );
-
-      const newNodePositions = newNodes.map((n: NodeSingular) => ({
-        id: n.id(),
-        x: n.position().x,
-        y: n.position().y
-      }));
-
-      // Check Y-coordinates for new nodes
-      const newYCoords = newNodePositions.map(p => p.y);
-      const allNewAtZero = newYCoords.every(y => y === 0);
-
-      // Check all Y-coordinates (old + new)
-      const allYCoords = positions.map(p => p.y);
-      const uniqueYAll = new Set(allYCoords);
-
-      // Check for overlaps
-      let overlapCount = 0;
-      const MINIMUM_DISTANCE = 30;
-
-      for (let i = 0; i < positions.length; i++) {
-        for (let j = i + 1; j < positions.length; j++) {
-          const dist = Math.hypot(
-            positions[i].x - positions[j].x,
-            positions[i].y - positions[j].y
-          );
-          if (dist < MINIMUM_DISTANCE) {
-            overlapCount++;
-          }
-        }
-      }
-
-      return {
-        totalNodes: nodes.length,
-        newNodePositions,
-        allNewAtZero,
-        uniqueYLevels: uniqueYAll.size,
-        overlapCount
-      };
-    });
-
-    console.log(`Final graph: ${finalState.totalNodes} nodes`);
-    console.log(`New nodes:`, finalState.newNodePositions);
-    console.log(`Unique Y levels: ${finalState.uniqueYLevels}`);
-    console.log(`Overlaps: ${finalState.overlapCount}`);
-
-    // Verify incremental nodes have proper positions
+    const finalState = await realFolder.getIncrementalLayoutState(appWindow);
     expect(finalState.totalNodes).toBe(bulkLoadState.nodeCount + 3);
     expect(finalState.newNodePositions.length).toBe(3);
-
-    // Check that new nodes have positions
-    // Note: Layout behavior may place nodes at origin initially before auto-layout runs
     const uniqueNewPositions = new Set(
       finalState.newNodePositions.map(p => `${Math.round(p.x)},${Math.round(p.y)}`)
     );
-    // At least verify nodes were created (position quality may vary based on layout timing)
     expect(uniqueNewPositions.size).toBeGreaterThanOrEqual(1);
-
-    console.log('✓ Incremental nodes positioned correctly');
-    console.log('✓ No excessive overlaps detected');
-
-    // Clean up the test files
     for (const file of newFiles) {
-      const filePath = path.join(FIXTURE_VAULT_PATH, file.name);
-      await fs.unlink(filePath);
+      await vault.deleteVaultFile(file.name);
     }
-
-    console.log('✓ Test files cleaned up');
-    console.log('✅ Bulk load + incremental layout test completed successfully!');
   });
 
   test('should scale node size and border width based on degree', async ({ appWindow }) => {
-    console.log('=== Testing node size scaling with degree ===');
-
-    // Verify app loaded properly (already waited in fixture)
-    const appReady = await appWindow.evaluate(() => {
-      return !!(window as ExtendedWindow).cytoscapeInstance &&
-             !!(window as ExtendedWindow).electronAPI;
-    });
+    const appReady = await realFolder.isAppReady(appWindow);
     expect(appReady).toBe(true);
-    console.log('✓ App loaded successfully');
-
-    // Wait for auto-load to complete (vault loads from config automatically)
-    await expect.poll(async () => {
-      return appWindow.evaluate(() => {
-        const cy = (window as ExtendedWindow).cytoscapeInstance;
-        if (!cy) return 0;
-        return cy.nodes().length;
-      });
-    }, {
+    await expect.poll(() => realFolder.getNodeCount(appWindow), {
       message: 'Waiting for graph to load nodes',
       timeout: 15000,
       intervals: [500, 1000, 1000]
     }).toBeGreaterThan(0);
-
-    // Get nodes with their degrees and dimensions
-    // Note: manually calculate degree if not set by updateNodeDegrees()
-    const nodeSizeData = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not initialized');
-
-      // Filter out nodes with special classes that override border-width styling
-      // These classes (hover, pinned, editor-shadow, terminal-shadow) set fixed border-width values
-      // that don't scale with degree, so they would skew our test results
-      const normalNodes = cy.nodes().filter((n: NodeSingular) => {
-        return n.id().includes(".")
-      });
-
-      const nodeData = normalNodes.map((n: NodeSingular) => {
-        // Check if degree is already set, otherwise calculate it
-        let degree = n.data('degree');
-        if (degree === undefined || degree === null) {
-          degree = n.degree(); // Calculate from actual connections
-          n.data('degree', degree); // Set it for styling
-        }
-
-        return {
-          id: n.id(),
-          label: n.data('label'),
-          degree: degree,
-          width: n.width(),
-          height: n.height(),
-          borderWidth: parseFloat(n.style('border-width'))
-        };
-      });
-
-      // Sort by degree to get high and low degree nodes
-      nodeData.sort((a, b) => a.degree - b.degree);
-
-      return {
-        all: nodeData,
-        lowest: nodeData[0],
-        highest: nodeData[nodeData.length - 1]
-      };
-    });
-
-    console.log('GraphNode with lowest degree:', nodeSizeData.lowest);
-    console.log('GraphNode with highest degree:', nodeSizeData.highest);
-
-    // Verify degree data is set
+    const nodeSizeData = await realFolder.getNodeSizeData(appWindow);
     expect(nodeSizeData.lowest.degree).toBeGreaterThanOrEqual(0);
     expect(nodeSizeData.highest.degree).toBeGreaterThan(nodeSizeData.lowest.degree);
-
-    // Verify size scaling: higher degree -> larger dimensions
     expect(nodeSizeData.highest.width).toBeGreaterThan(nodeSizeData.lowest.width);
     expect(nodeSizeData.highest.height).toBeGreaterThan(nodeSizeData.lowest.height);
-
-    // Log border width scaling behavior - border width may not always scale with degree
-    // depending on the specific nodes in the test fixture and layout timing
-    const borderWidthDiff = nodeSizeData.highest.borderWidth - nodeSizeData.lowest.borderWidth;
-    console.log(`Border width diff: ${borderWidthDiff} (highest - lowest)`);
-    // Just verify we got valid border width values, don't assert on the relationship
-    // since this can vary based on fixture state
     expect(nodeSizeData.highest.borderWidth).toBeGreaterThan(0);
     expect(nodeSizeData.lowest.borderWidth).toBeGreaterThan(0);
-
-    console.log('✓ GraphNode size scales correctly with degree');
-    console.log(`  Low degree (${nodeSizeData.lowest.degree}): ${Math.round(nodeSizeData.lowest.width)}x${Math.round(nodeSizeData.lowest.height)}px, border: ${nodeSizeData.lowest.borderWidth}px`);
-    console.log(`  High degree (${nodeSizeData.highest.degree}): ${Math.round(nodeSizeData.highest.width)}x${Math.round(nodeSizeData.highest.height)}px, border: ${nodeSizeData.highest.borderWidth}px`);
-
-    // Take screenshot for visual verification
     await appWindow.screenshot({ path: 'e2e-tests/test-results/degree-scaling-visualization.png' });
-    console.log('✓ Screenshot saved to e2e-tests/test-results/degree-scaling-visualization.png');
-
-    console.log('✓ GraphNode degree scaling test completed');
   });
 
   test('should select multiple nodes via box selection', async ({ appWindow }) => {
-    console.log('=== Testing box selection ===');
-
-    // Wait for auto-load to complete (vault loads from config automatically)
-    await expect.poll(async () => {
-      return appWindow.evaluate(() => {
-        const cy = (window as ExtendedWindow).cytoscapeInstance;
-        if (!cy) return 0;
-        return cy.nodes().length;
-      });
-    }, {
+    await expect.poll(() => realFolder.getNodeCount(appWindow), {
       message: 'Waiting for graph to load nodes',
       timeout: 15000,
       intervals: [500, 1000, 1000]
     }).toBeGreaterThan(0);
-
-    // Test box selection by programmatically selecting nodes
-    // This e2e-tests that boxSelectionEnabled is set and the boxend event fires
-    const selectionResult = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not initialized');
-
-      // Get all nodes
-      const allNodes = cy.nodes();
-      console.log(`[Test] Found ${allNodes.length} nodes in graph`);
-
-      if (allNodes.length < 3) {
-        throw new Error('Need at least 3 nodes for box selection test');
-      }
-
-      // Select first 3 nodes programmatically to simulate box selection result
-      const nodesToSelect = allNodes.slice(0, 3);
-      nodesToSelect.forEach((n: NodeSingular) => { n.select(); });
-
-      // Trigger boxend event manually to test the event handler
-      cy.trigger('boxend');
-
-      // Get selected nodes
-      const selected = cy.$('node:selected');
-
-      return {
-        totalNodes: allNodes.length,
-        selectedCount: selected.length,
-        selectedIds: selected.map((n: NodeSingular) => n.id()),
-        selectedLabels: selected.map((n: NodeSingular) => n.data('label'))
-      };
-    });
-
-    console.log(`Total nodes: ${selectionResult.totalNodes}`);
-    console.log(`Selected ${selectionResult.selectedCount} nodes:`, selectionResult.selectedLabels);
-
-    // Verify nodes were selected
+    const selectionResult = await realFolder.selectFirstThreeNodesForBoxSelection(appWindow);
     expect(selectionResult.selectedCount).toBe(3);
     expect(selectionResult.selectedIds.length).toBe(3);
-    console.log('✓ Box selection worked: 3 nodes selected');
-    console.log('✓ boxend event handler fired (check console logs)');
-
-    // Deselect all nodes
-    const deselectedCount = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) return -1;
-
-      cy.nodes().unselect();
-      return cy.$('node:selected').length;
-    });
+    const deselectedCount = await realFolder.deselectAllNodes(appWindow);
 
     expect(deselectedCount).toBe(0);
-    console.log('✓ Deselection worked');
-    console.log('✓ Box selection test completed');
   });
 
   test.skip('should add node via right-click context menu at graph position and open editor with file sync', async ({ appWindow }) => {
-    console.log('=== Testing Right-Click Add GraphNode with Editor and File Sync (BEHAVIORAL TEST) ===');
-    console.log('This test simulates the full right-click workflow: node creation + positioning + editor opening + file sync');
-
-    // Wait for auto-load to complete (vault loads from config automatically)
-    await expect.poll(async () => {
-      return appWindow.evaluate(() => {
-        const cy = (window as ExtendedWindow).cytoscapeInstance;
-        if (!cy) return 0;
-        return cy.nodes().length;
-      });
-    }, {
+    await expect.poll(() => realFolder.getNodeCount(appWindow), {
       message: 'Waiting for graph to load nodes',
       timeout: 15000,
       intervals: [500, 1000, 1000]
     }).toBeGreaterThan(0);
-
-    console.log('=== Step 1: Get initial node count ===');
-    const initialCount = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not initialized');
-      return cy.nodes().length;
-    });
-    console.log(`Initial node count: ${initialCount}`);
-
-    console.log('=== Step 2: Choose target position for new node (empty area of graph) ===');
-    const clickPosition = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not initialized');
-
-      // Find an empty area in the graph
-      const nodes = cy.nodes();
-      let maxX = -1000;
-      let maxY = -1000;
-      nodes.forEach((n: NodeSingular) => {
-        const pos = n.position();
-        if (pos.x > maxX) maxX = pos.x;
-        if (pos.y > maxY) maxY = pos.y;
-      });
-
-      // Target position: 300px right and 200px down from max node position
-      return { x: maxX + 300, y: maxY + 200 };
-    });
-    console.log(`Target position for new node: (${clickPosition.x}, ${clickPosition.y})`);
-
-    console.log('=== Step 3: Simulate right-click "Add GraphNode Here" action ===');
-    console.log('(Invoking testHelpers.addNodeAtPosition to trigger the full workflow)');
-
-    // Use the test helper to trigger the complete add-node-at-position workflow
-    // This includes: file creation + pending position storage + editor opening
-    await appWindow.evaluate(async (pos) => {
-      const w = (window as ExtendedWindow);
-
-      if (!w.testHelpers?.addNodeAtPosition) {
-        throw new Error('testHelpers.addNodeAtPosition not available');
-      }
-
-      console.log(`[Test] Calling testHelpers.addNodeAtPosition(${pos.x}, ${pos.y})`);
-      await w.testHelpers.addNodeAtPosition(pos);
-    }, clickPosition);
-
-    console.log('✓ Add node workflow triggered');
-
-    console.log('=== Step 4: Wait for new node to appear in graph ===');
-    await expect.poll(async () => {
-      return appWindow.evaluate(() => {
-        const cy = (window as ExtendedWindow).cytoscapeInstance;
-        if (!cy) return 0;
-        return cy.nodes().length;
-      });
-    }, {
+    const initialCount = await realFolder.getNodeCount(appWindow);
+    const clickPosition = await realFolder.getRightClickNodePosition(appWindow);
+    await realFolder.addNodeAtPosition(appWindow, clickPosition);
+    await expect.poll(() => realFolder.getNodeCount(appWindow), {
       message: 'Waiting for new node to be added to graph',
       timeout: 10000
     }).toBeGreaterThan(initialCount);
-
-    console.log('✓ New node added to graph');
-
-    console.log('=== Step 5: Verify node position is near click location ===');
-    const positionCheck = await appWindow.evaluate((clickPos) => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not initialized');
-
-      // Find all nodes, filter out the new one (likely has pattern _123 in ID)
-      const allNodes = cy.nodes();
-      const newNodes = allNodes.filter((n: NodeSingular) => {
-        const id = n.id();
-        // New nodes have IDs like "_123" or numeric pattern from standalone creation
-        return /^_?\d+$/.test(id);
-      });
-
-      if (newNodes.length === 0) {
-        return { success: false, message: 'No new node found', nodeId: null };
-      }
-
-      // Get the most recently added one (last in list)
-      const newNode = newNodes[newNodes.length - 1];
-      const nodePos = newNode.position();
-      const nodeId = newNode.id();
-
-      const distance = Math.sqrt(
-        Math.pow(nodePos.x - clickPos.x, 2) +
-        Math.pow(nodePos.y - clickPos.y, 2)
-      );
-
-      // Allow generous radius for layout adjustments
-      // NOTE: Auto-layout may reposition the node, so we use a generous threshold
-      // The key behavior is that the node is CREATED, not necessarily at the exact position
-      const maxDistance = 2000; // Very generous - just verify node was created
-      const success = distance <= maxDistance;
-
-      console.log(`[Test] New node ${nodeId} at (${nodePos.x.toFixed(1)}, ${nodePos.y.toFixed(1)}), ` +
-                 `click at (${clickPos.x}, ${clickPos.y}), distance: ${distance.toFixed(1)}px`);
-
-      return {
-        success,
-        message: `Node at (${nodePos.x.toFixed(1)}, ${nodePos.y.toFixed(1)}), distance: ${distance.toFixed(1)}px`,
-        nodeId: nodeId,
-        distance: distance
-      };
-    }, clickPosition);
+    const positionCheck = await realFolder.checkNewNodePosition(appWindow, clickPosition);
 
     expect(positionCheck.success).toBe(true);
     expect(positionCheck.nodeId).toBeTruthy();
     const newNodeId = positionCheck.nodeId!;
-    console.log(`✓ Node ${newNodeId} positioned within acceptable radius: ${positionCheck.message}`);
-
-    console.log('=== Step 6: Verify markdown editor opened automatically ===');
     const editorId = `editor-${newNodeId}`;
-    await expect.poll(async () => {
-      return appWindow.evaluate((edId) => {
-        const cy = (window as ExtendedWindow).cytoscapeInstance;
-        if (!cy) return false;
-        const shadowNode = cy.getElementById(edId);
-        return shadowNode.length > 0;
-      }, editorId);
-    }, {
+    await expect.poll(() => realFolder.hasGraphElement(appWindow, editorId), {
       message: `Waiting for editor ${editorId} to open`,
       timeout: 5000
     }).toBe(true);
-
-    console.log(`✓ Editor ${editorId} opened automatically`);
-
-    console.log('=== Step 7: Wait for CodeMirror editor to render ===');
     await appWindow.waitForSelector(`#window-${editorId} .cm-editor`, { timeout: 5000 });
-    console.log('✓ CodeMirror editor rendered');
-
-    console.log('=== Step 7b: Wait for editor to be fully initialized ===');
-    // Give the editor a moment to fully initialize
     await appWindow.waitForTimeout(500);
-    console.log('✓ Editor is ready');
-
-    console.log('=== Step 8: Edit content in markdown editor ===');
     const testContent = `---\nnode_id: ${newNodeId}\ntitle: Test Node ${newNodeId}\n---\n\n# Updated Content\n\nThis content was added by the E2E test to verify file sync.`;
 
-    await appWindow.evaluate((args) => {
-      const [edId, content] = args;
-      const w = (window as ExtendedWindow);
-      const editor = w.testHelpers?.getEditorInstance(edId);
-      if (!editor) {
-        throw new Error(`Editor instance not found for ${edId}`);
-      }
-
-      editor.setValue(content);
-      console.log(`[Test] Updated editor content for ${edId}`);
-    }, [editorId, testContent]);
-
-    console.log('✓ Editor content updated');
-
-    console.log('=== Step 8b: Verify editor value actually changed ===');
-    const editorValue = await appWindow.evaluate((edId) => {
-      const w = (window as ExtendedWindow);
-      const editor = w.testHelpers?.getEditorInstance(edId);
-      return editor?.getValue() ?? null;
-    }, editorId);
-    console.log(`Editor value length: ${editorValue?.length ?? 0}`);
-    console.log(`Expected content contains "Updated Content": ${editorValue?.includes('Updated Content') ?? false}`);
+    await realFolder.setEditorContent(appWindow, editorId, testContent);
+    const editorValue = await realFolder.getEditorValue(appWindow, editorId);
     expect(editorValue).toContain('Updated Content');
-    console.log('✓ Editor value successfully changed');
-
-    console.log('=== Step 9: Wait for auto-save to write to file system ===');
-
-    console.log('=== Step 10: Verify file content matches editor content ===');
-    // Find the specific new markdown file for the node we created
-    // newNodeId already contains the underscore prefix (e.g. "_8")
     const expectedFileName = `${newNodeId}.md`;
-    const filePath = path.join(FIXTURE_VAULT_PATH, expectedFileName);
-
-    // Poll for file existence with timeout (auto-save can take some time)
-    await test.expect.poll(async () => {
-      try {
-        await fs.access(filePath);
-        return true;
-      } catch {
-        return false;
-      }
-    }, {
+    const filePath = vault.filePathInVault(expectedFileName);
+    await test.expect.poll(() => vault.fileExists(filePath), {
       message: `Waiting for file ${expectedFileName} to be created`,
       timeout: 5000,
       intervals: [100, 200, 500] // Check frequently at first, then less often
     }).toBe(true);
 
-    console.log(`✓ Found new file: ${expectedFileName}`);
-
-    const fileContent = await fs.readFile(filePath, 'utf-8');
+    const fileContent = await vault.readTextFile(filePath);
 
     expect(fileContent).toContain('Updated Content');
     expect(fileContent).toContain('This content was added by the E2E test');
-    console.log('✓ File content matches editor updates');
-
-    console.log('=== Cleanup: Delete test file ===');
-    await fs.unlink(filePath);
-    console.log(`✓ Deleted test file: ${expectedFileName}`);
-
-    console.log('✓ Right-click add node with editor and file sync test completed');
+    await vault.deleteFilePath(filePath);
   });
 
   test('should open search with cmd-f and navigate to selected node', async ({ appWindow }) => {
-    console.log('\n=== Starting ninja-keys search navigation test ===');
-
-    console.log('=== Step 1: Wait for auto-load to complete ===');
-    // Wait for auto-load to complete (vault loads from config automatically)
-    await expect.poll(async () => {
-      return appWindow.evaluate(() => {
-        const cy = (window as ExtendedWindow).cytoscapeInstance;
-        if (!cy) return 0;
-        return cy.nodes().length;
-      });
-    }, {
+    await expect.poll(() => realFolder.getNodeCount(appWindow), {
       message: 'Waiting for graph to load nodes',
       timeout: 15000,
       intervals: [500, 1000, 1000]
     }).toBeGreaterThan(0);
 
-    const graphState = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not initialized');
-      return {
-        nodeCount: cy.nodes().length,
-        nodeLabels: cy.nodes().map((n: NodeSingular) => n.data('label') ?? n.id()).slice(0, 5)
-      };
-    });
+    const graphState = await realFolder.getGraphSummary(appWindow);
 
     expect(graphState.nodeCount).toBeGreaterThan(0);
-    console.log(`✓ Graph loaded with ${graphState.nodeCount} nodes`);
-    console.log(`  Sample nodes: ${graphState.nodeLabels.join(', ')}`);
-
-    console.log('=== Step 2: Get initial zoom/pan state ===');
-    const initialState = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not initialized');
-      const zoom = cy.zoom();
-      const pan = cy.pan();
-      return { zoom, pan };
-    });
-    console.log(`  Initial zoom: ${initialState.zoom}, pan: (${initialState.pan.x}, ${initialState.pan.y})`);
-
-    console.log('=== Step 3: Open ninja-keys search with keyboard shortcut ===');
-    // Simulate cmd-f (Meta+f on Mac, Ctrl+f elsewhere)
+    const initialState = await realFolder.getViewportState(appWindow);
     await appWindow.keyboard.press(process.platform === 'darwin' ? 'Meta+f' : 'Control+f');
-
-    // Wait for ninja-keys modal to appear
     await appWindow.waitForTimeout(300);
 
-    const ninjaKeysVisible = await appWindow.evaluate(() => {
-      const ninjaKeys = document.querySelector('ninja-keys');
-      if (!ninjaKeys) return false;
-      const shadowRoot = ninjaKeys.shadowRoot;
-      if (!shadowRoot) return false;
-      const modal = shadowRoot.querySelector('.modal');
-      // Check if modal exists and is not hidden
-      return modal !== null;
-    });
+    const ninjaKeysVisible = await realFolder.isNinjaKeysVisible(appWindow);
 
     expect(ninjaKeysVisible).toBe(true);
-    console.log('✓ ninja-keys search modal opened');
-
-    console.log('=== Step 4: Get a target node to search for ===');
-    const targetNode = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not initialized');
-      const nodes = cy.nodes();
-      if (nodes.length === 0) throw new Error('No nodes available');
-      // Get first node
-      const node = nodes[0];
-      return {
-        id: node.id(),
-        label: node.data('label') ?? node.id()
-      };
-    });
-
-    console.log(`  Target node: ${targetNode.label} (${targetNode.id})`);
-
-    console.log('=== Step 5: Type search query into ninja-keys ===');
-    // Type a few characters from the node label
+    const targetNode = await realFolder.getFirstSearchTargetNode(appWindow);
     const searchQuery = targetNode.label.substring(0, Math.min(5, targetNode.label.length));
     await appWindow.keyboard.type(searchQuery);
-
-    // Wait for search results to update
     await appWindow.waitForTimeout(300);
-    console.log(`  Typed search query: "${searchQuery}"`);
-
-    console.log('=== Step 6: Select first result with Enter ===');
     await appWindow.keyboard.press('Enter');
-
-    // Wait for navigation animation and fit to complete
     await appWindow.waitForTimeout(1000);
-
-    console.log('=== Step 7: Verify zoom/pan changed (node was fitted) ===');
-    const finalState = await appWindow.evaluate(() => {
-      const cy = (window as ExtendedWindow).cytoscapeInstance;
-      if (!cy) throw new Error('Cytoscape not initialized');
-      const zoom = cy.zoom();
-      const pan = cy.pan();
-      return { zoom, pan };
-    });
-
-    console.log(`  Final zoom: ${finalState.zoom}, pan: (${finalState.pan.x}, ${finalState.pan.y})`);
-
-    // Check that EITHER zoom or pan changed (cy.fit modifies these)
+    const finalState = await realFolder.getViewportState(appWindow);
     const zoomChanged = Math.abs(finalState.zoom - initialState.zoom) > 0.01;
     const panChanged = Math.abs(finalState.pan.x - initialState.pan.x) > 1 ||
                        Math.abs(finalState.pan.y - initialState.pan.y) > 1;
 
     expect(zoomChanged || panChanged).toBe(true);
-    console.log('✓ Graph viewport changed - node was fitted');
-
-    console.log('=== Step 8: Verify ninja-keys modal closed ===');
-    const ninjaKeysClosed = await appWindow.evaluate(() => {
-      const ninjaKeys = document.querySelector('ninja-keys');
-      if (!ninjaKeys) return true; // Not found means closed
-      const shadowRoot = ninjaKeys.shadowRoot;
-      if (!shadowRoot) return true;
-      const modal = shadowRoot.querySelector('.modal');
-      // Modal should be hidden or removed
-      if (!modal) return true;
-      const overlay = shadowRoot.querySelector('.modal-overlay');
-      // Check if overlay is visible (indicates open state)
-      return overlay ? getComputedStyle(overlay).display === 'none' : true;
-    });
+    const ninjaKeysClosed = await realFolder.isNinjaKeysClosed(appWindow);
 
     expect(ninjaKeysClosed).toBe(true);
-    console.log('✓ ninja-keys modal closed after selection');
-
-    console.log('✓ ninja-keys search navigation test completed');
   });
 
   test('should handle multiple consecutive cmd-f searches without focus issues', async ({ appWindow }) => {
-    console.log('\n=== Starting multiple consecutive search test ===');
-
-    console.log('=== Step 1: Wait for auto-load to complete ===');
-    // Wait for auto-load to complete (vault loads from config automatically)
-    await expect.poll(async () => {
-      return appWindow.evaluate(() => {
-        const cy = (window as ExtendedWindow).cytoscapeInstance;
-        if (!cy) return 0;
-        return cy.nodes().length;
-      });
-    }, {
+    await expect.poll(() => realFolder.getNodeCount(appWindow), {
       message: 'Waiting for graph to load nodes',
       timeout: 15000,
       intervals: [500, 1000, 1000]
     }).toBeGreaterThan(0);
-
-    console.log('=== Step 2: Get three different search queries ===');
-    // Use simple search terms that will match nodes in the fixture
-    // The test verifies that consecutive searches work, not that specific nodes are found
     const searchQueries = ['Agent', 'Test', 'Cloud'];
-
-    // Track viewport between iterations to verify each search causes navigation
-    let previousViewport: { zoom: number; pan: { x: number; y: number } } | null = null;
+    let previousViewport: ViewportState | null = null;
 
     for (let i = 0; i < 3; i++) {
       const searchQuery = searchQueries[i];
-      console.log(`\n=== Iteration ${i + 1}: Search using query "${searchQuery}" ===`);
-
-      console.log(`  Step ${i + 1}.1: Open search with Cmd-F`);
       await appWindow.keyboard.press(process.platform === 'darwin' ? 'Meta+f' : 'Control+f');
       await appWindow.waitForTimeout(300);
-
-      console.log(`  Step ${i + 1}.2: Verify modal opened`);
-      const modalOpen = await appWindow.evaluate(() => {
-        const ninjaKeys = document.querySelector('ninja-keys');
-        return ninjaKeys?.shadowRoot?.querySelector('.modal') !== null;
-      });
+      const modalOpen = await realFolder.isNinjaKeysModalOpen(appWindow);
 
       if (!modalOpen) {
         throw new Error(`Search modal failed to open on iteration ${i + 1}`);
       }
-      console.log(`  ✓ Modal opened successfully on iteration ${i + 1}`);
-
-      console.log(`  Step ${i + 1}.3: Type search query`);
-      // Use unique search query that will match only the target node
       await appWindow.keyboard.type(searchQuery);
       await appWindow.waitForTimeout(300); // Give search time to filter results
-
-      console.log(`  Step ${i + 1}.4: Select result with Enter`);
       await appWindow.keyboard.press('Enter');
       await appWindow.waitForTimeout(1000); // Wait for fit animation to complete
-
-      console.log(`  Step ${i + 1}.5: Verify viewport changed (navigation occurred)`);
-      // Check that the viewport changed - this indicates navigation happened
-      // Note: We don't check which specific node was selected because the handler
-      // may navigate without leaving the node in selected state
-      const currentViewport = await appWindow.evaluate(() => {
-        const cy = (window as ExtendedWindow).cytoscapeInstance;
-        if (!cy) throw new Error('Cytoscape not initialized');
-        const zoom = cy.zoom();
-        const pan = cy.pan();
-        return { zoom, pan };
-      });
-
-      // For first iteration, just verify a viewport exists
+      const currentViewport = await realFolder.getViewportState(appWindow);
       if (i === 0) {
         expect(currentViewport.zoom).toBeGreaterThan(0);
-        console.log(`  ✓ Navigation occurred on iteration ${i + 1} (viewport established)`);
       } else {
-        // For subsequent iterations, verify viewport changed from previous
         const viewportChanged = previousViewport !== null && (
           Math.abs(currentViewport.zoom - previousViewport.zoom) > 0.01 ||
           Math.abs(currentViewport.pan.x - previousViewport.pan.x) > 1 ||
           Math.abs(currentViewport.pan.y - previousViewport.pan.y) > 1
         );
         expect(viewportChanged).toBe(true);
-        console.log(`  ✓ Navigation occurred on iteration ${i + 1} (viewport changed)`);
       }
-
-      // Store this viewport for next iteration comparison
       previousViewport = currentViewport;
-
-      console.log(`  Step ${i + 1}.6: Verify modal closed`);
-      const modalClosed = await appWindow.evaluate(() => {
-        const ninjaKeys = document.querySelector('ninja-keys');
-        if (!ninjaKeys?.shadowRoot) return true;
-        const overlay = ninjaKeys.shadowRoot.querySelector('.modal-overlay');
-        return !overlay || getComputedStyle(overlay).display === 'none';
-      });
+      const modalClosed = await realFolder.isNinjaKeysModalClosed(appWindow);
 
       expect(modalClosed).toBe(true);
-      console.log(`  ✓ Modal closed after iteration ${i + 1}`);
-
-      // Brief pause between iterations
       await appWindow.waitForTimeout(200);
     }
-
-    console.log('\n✓ Successfully completed 3 consecutive searches without focus issues');
   });
 });
-
-export { test };
