+++ webapp/e2e-tests/electron/for_feature_development_not_LT_verification/graph_folder_loading/electron-real-folder/ui-helpers.ts
@@ -1,0 +1,226 @@
+import type { Page } from '@playwright/test';
+import type { NodeSingular } from 'cytoscape';
+import type {
+  ExtendedWindow,
+  GraphPosition,
+  GraphSummary,
+  NodePositionCheck,
+  SearchTargetNode,
+  SelectionResult,
+  ViewportState
+} from './types';
+
+export const selectFirstThreeNodesForBoxSelection = (appWindow: Page): Promise<SelectionResult> =>
+  appWindow.evaluate(() => {
+    const cy = (window as ExtendedWindow).cytoscapeInstance;
+    if (!cy) throw new Error('Cytoscape not initialized');
+
+    // Get all nodes
+    const allNodes = cy.nodes();
+    console.log(`[Test] Found ${allNodes.length} nodes in graph`);
+
+    if (allNodes.length < 3) {
+      throw new Error('Need at least 3 nodes for box selection test');
+    }
+
+    // Select first 3 nodes programmatically to simulate box selection result
+    const nodesToSelect = allNodes.slice(0, 3);
+    nodesToSelect.forEach((n: NodeSingular) => { n.select(); });
+
+    // Trigger boxend event manually to test the event handler
+    cy.trigger('boxend');
+
+    // Get selected nodes
+    const selected = cy.$('node:selected');
+
+    return {
+      totalNodes: allNodes.length,
+      selectedCount: selected.length,
+      selectedIds: selected.map((n: NodeSingular) => n.id()),
+      selectedLabels: selected.map((n: NodeSingular) => n.data('label'))
+    };
+  });
+
+export const deselectAllNodes = (appWindow: Page): Promise<number> =>
+  appWindow.evaluate(() => {
+    const cy = (window as ExtendedWindow).cytoscapeInstance;
+    if (!cy) return -1;
+
+    cy.nodes().unselect();
+    return cy.$('node:selected').length;
+  });
+
+export const getRightClickNodePosition = (appWindow: Page): Promise<GraphPosition> =>
+  appWindow.evaluate(() => {
+    const cy = (window as ExtendedWindow).cytoscapeInstance;
+    if (!cy) throw new Error('Cytoscape not initialized');
+
+    // Find an empty area in the graph
+    const nodes = cy.nodes();
+    let maxX = -1000;
+    let maxY = -1000;
+    nodes.forEach((n: NodeSingular) => {
+      const pos = n.position();
+      if (pos.x > maxX) maxX = pos.x;
+      if (pos.y > maxY) maxY = pos.y;
+    });
+
+    // Target position: 300px right and 200px down from max node position
+    return { x: maxX + 300, y: maxY + 200 };
+  });
+
+export const addNodeAtPosition = (appWindow: Page, position: GraphPosition): Promise<void> =>
+  appWindow.evaluate(async (pos) => {
+    const w = (window as ExtendedWindow);
+
+    if (!w.testHelpers?.addNodeAtPosition) {
+      throw new Error('testHelpers.addNodeAtPosition not available');
+    }
+
+    console.log(`[Test] Calling testHelpers.addNodeAtPosition(${pos.x}, ${pos.y})`);
+    await w.testHelpers.addNodeAtPosition(pos);
+  }, position);
+
+export const checkNewNodePosition = (appWindow: Page, clickPosition: GraphPosition): Promise<NodePositionCheck> =>
+  appWindow.evaluate((clickPos) => {
+    const cy = (window as ExtendedWindow).cytoscapeInstance;
+    if (!cy) throw new Error('Cytoscape not initialized');
+
+    // Find all nodes, filter out the new one (likely has pattern _123 in ID)
+    const allNodes = cy.nodes();
+    const newNodes = allNodes.filter((n: NodeSingular) => {
+      const id = n.id();
+      // New nodes have IDs like "_123" or numeric pattern from standalone creation
+      return /^_?\d+$/.test(id);
+    });
+
+    if (newNodes.length === 0) {
+      return { success: false, message: 'No new node found', nodeId: null };
+    }
+
+    // Get the most recently added one (last in list)
+    const newNode = newNodes[newNodes.length - 1];
+    const nodePos = newNode.position();
+    const nodeId = newNode.id();
+
+    const distance = Math.sqrt(
+      Math.pow(nodePos.x - clickPos.x, 2) +
+      Math.pow(nodePos.y - clickPos.y, 2)
+    );
+
+    // Allow generous radius for layout adjustments
+    // NOTE: Auto-layout may reposition the node, so we use a generous threshold
+    // The key behavior is that the node is CREATED, not necessarily at the exact position
+    const maxDistance = 2000; // Very generous - just verify node was created
+    const success = distance <= maxDistance;
+
+    console.log(`[Test] New node ${nodeId} at (${nodePos.x.toFixed(1)}, ${nodePos.y.toFixed(1)}), ` +
+               `click at (${clickPos.x}, ${clickPos.y}), distance: ${distance.toFixed(1)}px`);
+
+    return {
+      success,
+      message: `Node at (${nodePos.x.toFixed(1)}, ${nodePos.y.toFixed(1)}), distance: ${distance.toFixed(1)}px`,
+      nodeId: nodeId,
+      distance: distance
+    };
+  }, clickPosition);
+
+export const hasGraphElement = (appWindow: Page, elementId: string): Promise<boolean> =>
+  appWindow.evaluate((id) => {
+    const cy = (window as ExtendedWindow).cytoscapeInstance;
+    if (!cy) return false;
+    const shadowNode = cy.getElementById(id);
+    return shadowNode.length > 0;
+  }, elementId);
+
+export const setEditorContent = (appWindow: Page, editorId: string, content: string): Promise<void> =>
+  appWindow.evaluate((args) => {
+    const [edId, editorContent] = args;
+    const w = (window as ExtendedWindow);
+    const editor = w.testHelpers?.getEditorInstance(edId);
+    if (!editor) {
+      throw new Error(`Editor instance not found for ${edId}`);
+    }
+
+    editor.setValue(editorContent);
+    console.log(`[Test] Updated editor content for ${edId}`);
+  }, [editorId, content] as [string, string]);
+
+export const getEditorValue = (appWindow: Page, editorId: string): Promise<string | null> =>
+  appWindow.evaluate((edId) => {
+    const w = (window as ExtendedWindow);
+    const editor = w.testHelpers?.getEditorInstance(edId);
+    return editor?.getValue() ?? null;
+  }, editorId);
+
+export const getGraphSummary = (appWindow: Page): Promise<GraphSummary> =>
+  appWindow.evaluate(() => {
+    const cy = (window as ExtendedWindow).cytoscapeInstance;
+    if (!cy) throw new Error('Cytoscape not initialized');
+    return {
+      nodeCount: cy.nodes().length,
+      nodeLabels: cy.nodes().map((n: NodeSingular) => n.data('label') ?? n.id()).slice(0, 5)
+    };
+  });
+
+export const getViewportState = (appWindow: Page): Promise<ViewportState> =>
+  appWindow.evaluate(() => {
+    const cy = (window as ExtendedWindow).cytoscapeInstance;
+    if (!cy) throw new Error('Cytoscape not initialized');
+    const zoom = cy.zoom();
+    const pan = cy.pan();
+    return { zoom, pan };
+  });
+
+export const isNinjaKeysVisible = (appWindow: Page): Promise<boolean> =>
+  appWindow.evaluate(() => {
+    const ninjaKeys = document.querySelector('ninja-keys');
+    if (!ninjaKeys) return false;
+    const shadowRoot = ninjaKeys.shadowRoot;
+    if (!shadowRoot) return false;
+    const modal = shadowRoot.querySelector('.modal');
+    // Check if modal exists and is not hidden
+    return modal !== null;
+  });
+
+export const getFirstSearchTargetNode = (appWindow: Page): Promise<SearchTargetNode> =>
+  appWindow.evaluate(() => {
+    const cy = (window as ExtendedWindow).cytoscapeInstance;
+    if (!cy) throw new Error('Cytoscape not initialized');
+    const nodes = cy.nodes();
+    if (nodes.length === 0) throw new Error('No nodes available');
+    // Get first node
+    const node = nodes[0];
+    return {
+      id: node.id(),
+      label: node.data('label') ?? node.id()
+    };
+  });
+
+export const isNinjaKeysClosed = (appWindow: Page): Promise<boolean> =>
+  appWindow.evaluate(() => {
+    const ninjaKeys = document.querySelector('ninja-keys');
+    if (!ninjaKeys) return true; // Not found means closed
+    const shadowRoot = ninjaKeys.shadowRoot;
+    if (!shadowRoot) return true;
+    const modal = shadowRoot.querySelector('.modal');
+    // Modal should be hidden or removed
+    if (!modal) return true;
+    const overlay = shadowRoot.querySelector('.modal-overlay');
+    // Check if overlay is visible (indicates open state)
+    return overlay ? getComputedStyle(overlay).display === 'none' : true;
+  });
+
+export const isNinjaKeysModalOpen = (appWindow: Page): Promise<boolean> =>
+  appWindow.evaluate(() => {
+    const ninjaKeys = document.querySelector('ninja-keys');
+    return ninjaKeys?.shadowRoot?.querySelector('.modal') !== null;
+  });
+
+export const isNinjaKeysModalClosed = (appWindow: Page): Promise<boolean> =>
+  appWindow.evaluate(() => {
+    const ninjaKeys = document.querySelector('ninja-keys');
+    if (!ninjaKeys?.shadowRoot) return true;
+    const overlay = ninjaKeys.shadowRoot.querySelector('.modal-overlay');
+    return !overlay || getComputedStyle(overlay).display === 'none';
+  });
